requirements.txt
python-telegram-bot==20.7
httpx==0.27.2
requests==2.32.3
websockets==12.0
python-ulid==2.7.0
python-dotenv==1.0.1
groq==0.11.0
